Fix error handling in read/write and remove Cstruct.set_len
Cast result to int32 before multiplying by -1. Without this the result
is not negative and the error is not caught in the Ocaml code.

Also documents the minimum read buffer length (max_packet_size) and
replaces the deprecated call to Cstruct.set_len with Cstruct.sub.

+++ lib/vmnet_stubs.c
@@ -173,7 +173,7 @@ caml_vmnet_read(value v_vmnet, value v_ba, value v_ba_off, value v_ba_len)
   int pktcnt = 1;
   vmnet_return_t res = vmnet_read(iface, &v, &pktcnt);
   if (res != VMNET_SUCCESS)
-    CAMLreturn(Val_int((-1)*res));
+    CAMLreturn(Val_int((-1)*(int32_t)res));
   else if (pktcnt <= 0)
     CAMLreturn(Val_int(0));
   else
@@ -199,5 +199,5 @@ caml_vmnet_write(value v_vmnet, value v_ba, value v_ba_off, value v_ba_len)
   if (res == VMNET_SUCCESS)
     CAMLreturn(Val_int(v.vm_pkt_size));
   else
-    CAMLreturn(Val_int((-1)*res));
+    CAMLreturn(Val_int((-1)*(int32_t)res));
 }
